Move plugin system tests to model_library_test.py for CI discovery
The CI workflow explicitly runs specific test files in groups. Group 4
includes test/model_library_test.py but the new plugin_system_test.py
was not included in any group. By moving the tests to model_library_test.py,
they will now be picked up by the CI test runner and coverage reports.

This ensures the 90+ plugin system tests are actually executed during CI,
providing proper coverage for the new plugin discovery functionality.
